added keybindings for several components, including the main panel, the MessageWindows, and the FolderWindow.
note that while the mainPanel's keybindings work just fine, the other windows'
don't.  that's because JComponent.WHEN_ANCESTOR_OF_FOCUSED_COMPONENT
doesn't seem to be doing what it should.  i suppose a workaround could be
to add it both to the windows themselves and the windows' child components,
but that seems wasteful.


git-svn-id: svn://svn.code.sf.net/p/pooka/code/trunk/pooka@135 cf05c91d-1d0e-0410-8b41-cf3121da992e