mass updates.
added the Attachment class (and MBPAttachment subclass).

the new paradigm goes like this:  a message gets its attachments parsed once
into a bundle.  this bundle wraps all of the datahandlers so that we don't
actually have to store any of the data itself.  then each request gets the
cached info instead of having to request it each time.


git-svn-id: svn://svn.code.sf.net/p/pooka/code/trunk/pooka@510 cf05c91d-1d0e-0410-8b41-cf3121da992e